requirements.txt
giallure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
glob2==0.7
importlib-metadata==1.3.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
parse==1.13.0
parse-type==0.5.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
pytest-bdd==3.2.1
pytest-ordering==0.6
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
